+++ src/lib/lisp/gc.c
@@ -255,12 +255,12 @@ check_xlat:
             xlat_full = true;
 #ifdef VERBOSE_GC
         outn (total_removed); outs ("B freed."); terpri ();
-#endif
+#endif // #ifdef VERBOSE_GC
     } while ((++heap)->start);
 #else // #ifdef FRAGMENTED_HEAP
     // Save free pointer.
     heap_free = d;
-#endif
+#endif // #ifdef FRAGMENTED_HEAP
 
 #ifndef NDEBUG
     bzero (d, heap_end - d);
@@ -362,8 +362,12 @@ restart:
 
     // Mark global pointers.
     for (gp = global_pointers; *gp; gp++)
+        {
+        outhw ((unsigned long) **gp); out (' ');
         mark (**gp);
+        }
 
+        outs ("stack"); terpri ();
     // Mark GC'ed stack.
     for (p = stack; p != stack_end; p += sizeof (lispptr))
         mark (*(lispptr *) p);